+++ BE/BE/Repositories/Interfaces/IProductsRepository.cs
@@ -7,10 +7,10 @@ namespace BE.Repositories.Interfaces
     public interface IProductsRepository
     {
         Task<IEnumerable<Products>> GetAllAsync();
-        Task<Products?> GetByIdAsync(int id);
+        Task<Products?> GetByIdAsync(long id);
         Task<Products> AddAsync(Products model);
-        Task<Products?> UpdateAsync(int id, Products model);
-        Task<bool> DeleteAsync(int id);
+        Task<Products?> UpdateAsync(long id, Products model);
+        Task<bool> DeleteAsync(long id);
 
         List<Products> GetByProviderId(long providerId);
